pyproject.toml
[project]
name = "assignment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "langid>=1.1.6",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.0",
    "pdfplumber>=0.11.7",
    "pymupdf4llm>=1.27.2.2",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.7",
    "torch>=2.6.0",
    "trafilatura>=2.0.0",
    "wordcloud>=1.9.6",
]
